handle errors in chat

+++ srcs/backend/src/chat/chat.gateway.ts
@@ -80,11 +80,13 @@ export class ChatGateway implements OnGatewayConnection, OnGatewayDisconnect {
 
   @SubscribeMessage('messages')
   async getMessages(socket: any, roomId: number) {
-	try {
-		const room: RoomEntity = await this.chatService.getRoomById(roomId);
-		const messages = await this.chatService.findMessagesForRoom(room, +socket.handshake.headers.userid);
-    	this.server.to(socket.id).emit('messages', messages);
-	} catch {};
+    try {
+      const room: RoomEntity = await this.chatService.getRoomById(roomId);
+      if (room) {
+        const messages = await this.chatService.findMessagesForRoom(room, +socket.handshake.headers.userid);
+        this.server.to(socket.id).emit('messages', messages);
+      }
+    } catch { };
   }
 
   @SubscribeMessage('paginate_rooms')
@@ -101,8 +103,11 @@ export class ChatGateway implements OnGatewayConnection, OnGatewayDisconnect {
   @SubscribeMessage('leave_room')
   async onLeaveRoom(socket: Socket, roomId: number) {
     const user = await this.UsersService.getUserById(+socket.handshake.headers.userid);
+    if (!user) { return; }
     const room = await this.chatService.getRoomById(roomId);
+    if (!room) { return; }
     const member = await this.chatService.getMemberByRoomAndUser(room, user);
+    if (!member) { return; }
 
     if (await this.chatService.removeMemberFromRoom(room, member) == "delete_room") {
       await this.emitRooms(+socket.handshake.headers.userid, socket.id);
@@ -116,8 +121,8 @@ export class ChatGateway implements OnGatewayConnection, OnGatewayDisconnect {
     }
     else {
       await this.emitRooms(+socket.handshake.headers.userid, socket.id);
-	  const rooms = await this.chatService.getAllMyRooms(+socket.handshake.headers.userid);
-	  this.server.to(socket.id).emit('all_my_rooms', rooms);
+      const rooms = await this.chatService.getAllMyRooms(+socket.handshake.headers.userid);
+      this.server.to(socket.id).emit('all_my_rooms', rooms);
       const members = await this.chatService.getMembersByRoom(room);
       for (const member of members) {
         this.server.to(member.socketId).emit('members_room', members);
@@ -129,11 +134,10 @@ export class ChatGateway implements OnGatewayConnection, OnGatewayDisconnect {
   async getMembersOfRoom(socket: Socket, roomId: number) {
     if (roomId != null) {
       const room = await this.chatService.getRoomById(roomId);
-	  if (room)
-	  {
-		const members = await this.chatService.getMembersByRoom(room);
-		this.server.to(socket.id).emit('members_room', members);
-	  }
+      if (room) {
+        const members = await this.chatService.getMembersByRoom(room);
+        this.server.to(socket.id).emit('members_room', members);
+      }
     }
   }
